Initial commit.
Rebuilds the Android Leanback v17 library and supporting libaries as
APKLibs and Jars which can be used with maven and eclipse.  AOSP only
provides AAR for these items in the Android L preview which does means
that Eclipse ADT users are left in the dark if they want to build using
these libaries.